// types/home.d.ts
export interface User {
  id:                     number;
  first_name:             string;
  last_name:              string;
  email:                  string;
  payment_mode:           string;
  user_type:              string;
  gender:                 string;
  country_code:           null;
  mobile:                 string;
  device_token:           string;
  device_id:              string;
  device_type:            string;
  login_by:               string;
  language:               string;
  qrcode_url:             null;
  referral_unique_id:     null;
  referal_count:          number;
  trial_ends_at:          null;
  avatar:                 null;
  social_unique_id:       null;
  latitude:               null;
  longitude:              null;
  stripe_cust_id:         null;
  wallet_balance:         number;
  otp:                    number;
  email_verified_at:      string;
  blocked:                string;
  deleted_at:             null;
  created_at:             string;
  updated_at:             string;
  currency:               string;
  sos:                    string;
  app_contact:            string;
  measurement:            string;
  cash:                   number;
  debit_machine:          number;
  voucher:                number;
  online:                 number;
  card:                   number;
  payumoney:              number;
  paypal:                 number;
  paypal_adaptive:        number;
  braintree:              number;
  paytm:                  number;
  stripe_secret_key:      string;
  stripe_publishable_key: string;
  stripe_currency:        null;
  payumoney_environment:  null;
  payumoney_key:          null;
  payumoney_salt:         null;
  payumoney_auth:         null;
  paypal_environment:     null;
  paypal_currency:        null;
  paypal_client_id:       null;
  paypal_client_secret:   null;
  braintree_environment:  null;
  braintree_merchant_id:  null;
  braintree_public_key:   null;
  braintree_private_key:  null;
  referral_count:         string;
  referral_amount:        string;
  referral_text:          string;
  referral_total_count:   string;
  referral_total_amount:  number;
  referral_total_text:    string;
  ride_otp:               number;
  ride_toll:              number;
  rating:                 number;
}